[fio extras] clk: imx: CAAM clock enabled at init
When the CAAM driver is not installed, the clock is gated; op-tee
accesses to the CAAM registers will generate a data abort. During
tests, the first register that is being accessed (and causing a crash)
would be the at offset 0x14 in the job ring.

Signed-off-by: Jorge Ramirez-Ortiz <jorge@foundries.io>
Signed-off-by: Ricardo Salveti <ricardo@foundries.io>
(cherry picked from commit 0f71304f2f4e67bd71255150b44df4c3c4503895)
Signed-off-by: Vanessa Maegima <vanessa.maegima@foundries.io>